La np_image n'est plus un tableau denp.float64 converti en np.uint8 car la tronquation empeche l'optimisation du fit de converger. On Converti le tableau que pour generer une image avec la colormap car indispensable pour appliquer une colormap

+++ controler/center_emission_controler.py
@@ -59,7 +59,7 @@ class CenterEmissionControler(metaclass=SingletonMeta):
         self.raw_image = get_active_camera().snapshot('SPOT_LASER')
         self.np_image = get_substracted_image(self.raw_image, get_active_harmonisation_data().background_image) 
         # On met à jour l'image de la camera
-        colored_image = cv2.applyColorMap(self.np_image, cv2.COLORMAP_TURBO)
+        colored_image = cv2.applyColorMap(self.np_image.astype(np.uint8), cv2.COLORMAP_TURBO)
         height, width = self.np_image.shape
         bytes_per_line = width
         # Convertir en QImage
