+++ controler/miror_position_controler.py
@@ -14,7 +14,7 @@ import cv2
 from model.calculations import get_gauss_fit_params, get_euclidian_distance
 import threading
 import time
-
+import numpy as np
 DELAY = 200
 
 class MirorPositionControler(metaclass=SingletonMeta):
@@ -54,7 +54,7 @@ class MirorPositionControler(metaclass=SingletonMeta):
 
         self.np_image = get_active_camera().snapshot('SPOT_LASER') 
         # On met à jour l'image de la camera
-        colored_image = cv2.applyColorMap(self.np_image, cv2.COLORMAP_TURBO)
+        colored_image = cv2.applyColorMap(self.np_image.astype(np.uint8), cv2.COLORMAP_TURBO)
         height, width = self.np_image.shape
         bytes_per_line = width
         # Convertir en QImage
